theme function sub-class instances
